공정장 페이지 오류 수정

+++ src/main/java/kr/co/itwillbs/solex/notice/NoticeMapper.java
@@ -23,7 +23,7 @@ public interface NoticeMapper {
 	int updateNotice(Map<String, Object> param);
 
 	//글 삭제
-	int deleteNotice(int notId);
+	int deleteNotice(Long notId);
 	
 	Map<String, Object> getEmployeeInfo(Long empId);
 }
